+++ src/theme/components/button.ts
@@ -8,13 +8,8 @@ const baseStyle = {
   cursor: "pointer",
   color: "ui.white",
   fontWeight: "button.default",
-  height: "10",
   justifyContent: "center",
   lineHeight: "1.5",
-  maxHeight: "2.5rem",
-  minHeight: { base: defaultElementSizes.mobileFieldHeight, md: "auto" },
-  py: "inset.narrow",
-  px: "inset.default",
   textDecoration: "none",
   wordWrap: "normal",
   svg: {
@@ -33,20 +28,48 @@ const baseStyle = {
     pointerEvents: "none",
   },
 };
-// Styles for different visual variants:
-// primary, secondary, link, pill, iconOnly, callout, searchbar, noBrand
-const primary = {
-  bg: "ui.link.primary",
-  height: "none",
-  fontSize: "button.default",
-  minWidth: "none",
+
+const generalSizeValues = (size = "medium", isPill = false) => {
+  const sizes = {
+    small: {
+      fontSize: "button.small",
+      height: "fit-content",
+      minHeight: "auto",
+      px: isPill ? "s" : "button.small.px",
+      py: isPill ? "xxxs" : "button.small.py",
+    },
+    medium: {
+      fontSize: "button.default",
+      height: isPill ? "fit-content" : undefined,
+      minHeight: isPill
+        ? "auto"
+        : { base: defaultElementSizes.mobileFieldHeight, md: "auto" },
+      px: isPill ? "m" : "button.medium.px",
+      py: isPill ? "xxxs" : "button.medium.py",
+    },
+    large: {
+      fontSize: "button.large",
+      height: "fit-content",
+      minHeight: "auto",
+      px: isPill ? "l" : "button.large.px",
+      py: isPill ? "xxs" : "button.large.py",
+    },
+  };
+  return sizes[size];
 };
-const secondary = {
+
+// Styles for different visual variants:
+// primary, secondary, link, text, pill, iconOnly, callout, noBrand
+const primary = ({ buttonSize = "medium" }) => ({
+  bg: "ui.link.primary",
+  ...generalSizeValues(buttonSize),
+});
+const secondary = ({ buttonSize = "medium" }) => ({
   bg: "ui.white",
   border: "1px solid",
   borderColor: "ui.link.primary",
   color: "ui.link.primary",
-  fontSize: "button.default",
+  ...generalSizeValues(buttonSize),
   _hover: {
     bg: "ui.bg.default",
     borderColor: "ui.link.secondary",
@@ -57,12 +80,14 @@ const secondary = {
     borderColor: "ui.disabled.primary",
     color: "ui.disabled.primary",
   },
-};
-const link = {
+});
+// The "link" type is deprecated but we still want to style
+// it correctly for backwards compatibility.
+const link = ({ buttonSize = "medium" }) => ({
   bg: "transparent",
   color: "ui.link.primary",
-  lineHeight: "2.5",
   textDecoration: "underline",
+  ...generalSizeValues(buttonSize),
   _disabled: {
     bg: "transparent",
   },
@@ -70,11 +95,11 @@ const link = {
     bg: "transparent",
     color: "ui.link.secondary",
   },
-};
-const text = {
+});
+const text = ({ buttonSize = "medium" }) => ({
   bg: "transparent",
   color: "ui.link.primary",
-  fontSize: "button.default",
+  ...generalSizeValues(buttonSize),
   _disabled: {
     bg: "transparent",
   },
@@ -82,16 +107,14 @@ const text = {
     bg: "transparent",
     color: "ui.link.secondary",
   },
-};
-const pill = {
+});
+const pill = ({ buttonSize = "medium" }) => ({
   bg: "ui.white",
   border: "1px solid",
   borderColor: "ui.gray.light-cool",
   borderRadius: "pill",
   color: "inherit",
-  fontSize: "button.default",
-  py: "inset.narrow",
-  px: "inset.wide",
+  ...generalSizeValues(buttonSize, true),
   _hover: {
     bg: "ui.gray.xx-light-cool",
     borderColor: "ui.gray.medium",
@@ -99,38 +122,39 @@ const pill = {
   _disabled: {
     bg: "ui.gray.x-light-cool",
   },
-};
-const iconOnly = {
+});
+const iconOnly = ({ buttonSize = "medium" }) => ({
   bg: "ui.white",
   border: "1px solid",
   borderColor: "ui.link.primary",
   color: "ui.link.primary",
+  ...generalSizeValues(buttonSize),
+  paddingInlineStart: "inset.narrow",
+  paddingInlineEnd: "inset.narrow",
   _hover: {
     bg: "ui.bg.default",
     borderColor: "ui.link.primary",
     color: "ui.link.primary",
   },
-  paddingInlineStart: "inset.narrow",
-  paddingInlineEnd: "inset.narrow",
-};
-const callout = {
+});
+const callout = ({ buttonSize = "medium" }) => ({
   bg: "brand.primary",
-  fontSize: "button.default",
+  ...generalSizeValues(buttonSize),
   _hover: {
     bg: "brand.secondary",
   },
   _active: {
     bg: "brand.secondary",
   },
-};
-const noBrand = {
-  ...primary,
+});
+const noBrand = ({ buttonSize = "medium" }) => ({
   bg: "ui.black",
   color: "ui.white",
+  ...generalSizeValues(buttonSize),
   _hover: {
     bg: "ui.gray.x-dark",
   },
-};
+});
 
 const Button = {
   baseStyle,
@@ -139,6 +163,7 @@ const Button = {
     primary,
     secondary,
     text,
+    // Deprecated type. Use "text" instead.
     link,
     pill,
     iconOnly,